do not export Maybe type to avoid name conflicts
Maybe should *not* be defined by SBML.jl, rather by Base or some other standard
package, but we need it nevertheless. Exporting it means that other packages
will have problems with their own redefinitions of Maybe because name
conflicts. So just do not export it, leave others to redefine it or manually
copy it out, using `const Maybe = SBML.Maybe`.