use the safer Object_createChild SBML methods.
This depends on https://github.com/sbmlteam/libsbml/pull/252 to be merged,
because some of the functions were actually not exposed to the C API.

Creating the members from the parent objects seems safer because all XML
metadata and versioning info gets automatically propagated. Also the code is a
bit shorter.