remove the mandatory ID from the SBML events
Closes #254

SBML says that IDs on events are actually not mandatory, this softens the data
structure so that the ID-less events are supported. The vector-of-pairs
implementation is chosen for backwards compatibility (iteration through a
vector of pairs should work precisely like iteration through the dictionary
that was there before). Notably, this might look like a breaking change _but_
it is not -- the main use case that is broken by this commit is indexing the
events by ID, which was actually broken already.